# Reduce Startup Latency by moving the init container to a regular container