--- a/cmd/aws-vpc-cni-init/main.go
+++ b/cmd/aws-vpc-cni-init/main.go
@@ -16,6 +16,7 @@ package main
 
 import (
 	"os"
+	"time"
 
 	"github.com/aws/amazon-vpc-cni-k8s/pkg/procsyswrapper"
 	"github.com/aws/amazon-vpc-cni-k8s/utils/cp"
@@ -179,17 +180,17 @@ func _main() int {
 		return 1
 	}
 
-	// TODO: In order to speed up pod launch time, VPC CNI init container is not a Kubernetes init container.
+	// In order to speed up pod launch time, VPC CNI init container is not a Kubernetes init container.
 	// The VPC CNI container blocks on the existence of vpcCniInitDonePath
-	//err = cp.TouchFile(vpcCniInitDonePath)
-	//if err != nil {
-	//	log.WithError(err).Errorf("Failed to set VPC CNI init done")
-	//	return 1
-	//}
+	err = cp.TouchFile(vpcCniInitDonePath)
+	if err != nil {
+		log.WithError(err).Errorf("Failed to set VPC CNI init done")
+		return 1
+	}
 
 	log.Infof("CNI init container done")
 
-	// TODO: Since VPC CNI init container is a real container, it never exits
-	// time.Sleep(time.Duration(1<<63 - 1))
+	// Since VPC CNI init container is a real container, it never exits
+	time.Sleep(time.Duration(1<<63 - 1))
 	return 0
 }