@jayanthvn 2022-11-11
Need to run https://github.com/aws/amazon-vpc-cni-k8s/blob/master/Makefile#L342 and copy the new manifests here - https://github.com/aws/amazon-vpc-cni-k8s/tree/master/config/master.

Self managed addons upgrades/downgrades will be reapplying the manifests so I don't see an issue there but we need to confirm on managed addons upgrade/downgrade.